backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
pillow
python-dotenv
psycopg2-binary
cloudinary
rembg
rembg[gpu]